crypto_scheduler/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
watchfiles==0.21.0
redis==5.0.1